Final Mandatory

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: hmesrar <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/12 16:25:05 by hmesrar           #+#    #+#             */
-/*   Updated: 2022/11/14 00:54:49 by hmesrar          ###   ########.fr       */
+/*   Updated: 2022/11/14 18:06:56 by hmesrar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int 	ft_strlen(char *s);
 int    	ft_putchar(char c);
 int		ft_putstr(char *s);
-int    	ft_putnbr_xX(unsigned long nbr, char *base);
+int    	ft_putnbr_xX(unsigned int nbr, char *base);
 int		ft_putnbr_id(int nbr, char *base);
 int    	ft_putnbr_u(unsigned int nbr, char *base);
 int    	ft_putnbr_p(unsigned long long nbr, char *base);
