+++ ft_putnbr_xX.c
@@ -6,16 +6,16 @@
 /*   By: hmesrar <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/13 20:05:18 by hmesrar           #+#    #+#             */
-/*   Updated: 2022/11/14 00:54:33 by hmesrar          ###   ########.fr       */
+/*   Updated: 2022/11/14 18:07:10 by hmesrar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int    ft_putnbr_xX(unsigned long nbr, char *base)
+int    ft_putnbr_xX(unsigned int nbr, char *base)
 {
     int len;
-    unsigned int base_size;
+    unsigned long base_size;
 
     len = 0;
     base_size = ft_strlen(base);
